requirements.txt
anthropic>=0.8.0
beautifulsoup4>=4.12.0
elevenlabs>=0.3.0
fake-useragent>=1.4.0
openai>=1.12.0
PyPDF2>=3.0.0
pydub>=0.25.1
python-dotenv>=1.0.0
requests>=2.31.0
retrying>=1.3.4
streamlit>=1.31.0
tqdm>=4.66.0
urllib